[rev] move base VtkElementViewModel to "Base" folder
