+++ VtkMvvm/ViewModels/BullseyeViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using System.Numerics;
 using Kitware.VTK;
 using VtkMvvm.Models;
+using VtkMvvm.ViewModels.Base;
 
 namespace VtkMvvm.ViewModels;
