+++ VtkMvvm/ViewModels/CrosshairViewModel.cs
@@ -2,6 +2,7 @@
 using System.Numerics;
 using Kitware.VTK;
 using VtkMvvm.Models;
+using VtkMvvm.ViewModels.Base;
 
 namespace VtkMvvm.ViewModels;
